# docs/submission/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SAHYOG Blockchain Intelligence",
    "description": "Complaint intake, replayable ingestion, tracing, attribution and VASP federation.",
    "version": "0.16.0"
  },
  "paths": {
...
    "/api/v1/cases/{case_id}/taint": {
      "post": {
        "tags": [
          "Investigation workflow"
        ],
        "summary": "Taint",
        "operationId": "taint_api_v1_cases__case_id__taint_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaintInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/{case_id}/taint/summary": {
      "get": {
        "tags": [
          "Investigation workflow"
        ],
        "summary": "Taint Summary",
        "operationId": "taint_summary_api_v1_cases__case_id__taint_summary_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/addresses/{chain}/{address}/taint": {
      "get": {
        "tags": [
          "Investigation workflow"
        ],
        "summary": "Address Taint",
        "operationId": "address_taint_api_v1_addresses__chain___address__taint_get",
        "parameters": [
          {
            "name": "chain",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chain"
            }
          },
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Address"
            }
          },
          {
            "name": "case_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/gang-cases/{gang_id}": {
      "get": {
        "tags": [
          "Investigation workflow"
        ],
        "summary": "Gang",
        "operationId": "gang_api_v1_gang_cases__gang_id__get",
        "parameters": [
          {
            "name": "gang_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Gang Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/gang-cases/{gang_id}/split-preview": {
      "post": {
        "tags": [
          "Investigation workflow"
        ],
        "summary": "Split",
        "operationId": "split_api_v1_gang_cases__gang_id__split_preview_post",
        "parameters": [
          {
            "name": "gang_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Gang Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SplitInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/gang-cases/{gang_id}/communities": {
      "get": {
        "tags": [
          "Investigation workflow"
        ],
        "summary": "Communities",
        "operationId": "communities_api_v1_gang_cases__gang_id__communities_get",
        "parameters": [
          {
            "name": "gang_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Gang Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/{case_id}/wallet-farms": {
      "get": {
        "tags": [
          "Investigation workflow"
        ],
        "summary": "Farms",
        "operationId": "farms_api_v1_cases__case_id__wallet_farms_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/wallet-farms/{farm_id}/approve": {
      "post": {
        "tags": [
          "Investigation workflow"
        ],
        "summary": "Approve Farm",
        "operationId": "approve_farm_api_v1_wallet_farms__farm_id__approve_post",
        "parameters": [
          {
            "name": "farm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Farm Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/{case_id}/similar-operators": {
      "get": {
        "tags": [
          "Investigation workflow"
        ],
        "summary": "Similar",
        "operationId": "similar_api_v1_cases__case_id__similar_operators_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/collisions/{collision_id}/open-room": {
      "post": {
        "tags": [
          "Investigation workflow"
        ],
        "summary": "Open Room",
        "operationId": "open_room_api_v1_collisions__collision_id__open_room_post",
        "parameters": [
          {
            "name": "collision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Collision Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/requests/{request_id}": {
      "get": {
        "tags": [
          "Investigation workflow"
        ],
        "summary": "Request Detail",
        "operationId": "request_detail_api_v1_requests__request_id__get",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SplitInput": {
        "properties": {
          "available_amount": {
            "type": "string",
            "pattern": "^[0-9]{1,38}$",
            "title": "Available Amount"
          }
        },
        "type": "object",
        "required": [
          "available_amount"
        ],
        "title": "SplitInput"
      },
...
      "TaintInput": {
        "properties": {
          "method": {
            "type": "string",
            "enum": [
              "haircut",
              "fifo",
              "poison"
            ],
            "title": "Method",
            "default": "haircut"
          },
          "cutoff": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Cutoff",
            "default": "0.02"
          }
        },
        "type": "object",
        "title": "TaintInput"
      },
...
    }
  }
}